requirements.txt
streamlit==1.35.0
boto3==1.34.2
sseclient-py==1.8.0
python-dotenv==1.0.0
pandas==2.1.4
tabulate==0.9.0
beautifulsoup4
matplotlib==3.9.0